movie/pyproject.toml
[project]
name = "forana-movie"
description = "Simple Movie API for Ana Gardi."
readme = "README.rst"
version = "0.1"
dependencies = [
    "fastapi",
    "sqlmodel",
    "fastapi-crudrouter",
    "fake.py",
    "uvicorn",
    "starlette-admin",
    "sqlalchemy<2.0",
]
authors = [
    {name = "Artur Barseghyan", email = "artur.barseghyan@gmail.com"},
]
license = {text = "MIT"}
[project.optional-dependencies]
all = ["forana[dev,test,docs]"]
dev = [
    "black",
    "detect-secrets",
    "doc8",
    "ipython",
    "isort",
    "mypy",
    "pip-tools",
    "ruff",
]
test = [
    "pytest",
    "pytest-cov",
    "httpx",
]
docs = [
]

[tool.coverage.run]
relative_files = true
omit = [
    "test_*.py",
    "alembic/**",
]

[tool.coverage.report]
skip_covered = false
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]

[tool.pytest.ini_options]
python_files = [
    "test_*.py",
]

addopts = [
    "--cov=api",
    "--cov=db",
    "--cov=models",
    "--cov-append",
    "--cov-report=html",
    "--cov-report=xml",
    "--cov-report=term",
    "--cov-report=annotate"
]
